+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # React Native Module Changelog
 
+## Version 19.1.0 - July 17, 2024
+Minor release that fixes enabling or disabling all Airship features using `FEATURES_ALL` and adds possibility to enable and disable `Feature.FeatureFlags`.
+
+### Changes
+- Fixed enabling or disabling features using `FEATURE_ALL`.
+- Added possibility to enable and disable `Feature.FeatureFlags` using the privacy manager.
+
 ## Version 19.0.0 - July 9, 2024
 Major release that updates the Android Airship SDK to 18.